# packages/server/api.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Minimal ping API server
tags:
  - name: game
    description: Endpoints related to game sessions, including starting a new game, submitting answers, and real-time updates.
  - name: leaderboard
    description: Endpoints for retrieving leaderboard data, including pagination support.
paths:
  /api/v1/game/start:
    post:
      tags:
        - game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartGameRequest'
      responses:
        '200':
          description: Game started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartGameResponse'
  /api/v1/game/continue:
    post:
      tags:
        - game
      responses:
        '200':
          description: Game continued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartGameResponse'
        '404':
          description: Game session not found
  /api/v1/game/submit:
    post:
      tags:
        - game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Answer submitted successfully, response contains the next question and updated game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitAnswerResponse'
        '404':
          description: Game session not found
  /api/v1/game/sse/{sessionId}:
    get:
      tags:
        - game
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
            x-go-type: string
      responses:
        '200':
          description: SSE stream for real-time game updates
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/leaderboard:
    get:
      tags:
        - leaderboard
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Leaderboard data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLeaderboardResponse'
components:
  schemas:
    StartGameRequest:
      type: object
      required:
        - itemsUsed
      properties:
        autoVoice:
          type: boolean
        itemsUsed:
          type: array
          items:
            type: string
            format: uuid
            x-go-type: string
    StartGameResponse:
      type: object
      required:
        - sessionId
        - scrambledWord
        - scrambledWordDefinition
        - token
        - remainingSeconds
        - autoVoice
      properties:
        sessionId:
          type: string
          format: uuid
          x-go-type: string
        scrambledWord:
          type: string
          example: 'lpeap'
        scrambledWordDefinition:
          type: string
          example: 'A fruit that is often red or green.'
        token:
          type: string
          format: uuid
          x-go-type: string
        remainingSeconds:
          type: integer
          format: int32
          example: 60
        autoVoice:
          type: boolean
    SubmitAnswerRequest:
      type: object
      required:
        - sessionId
        - answer
        - token
      properties:
        sessionId:
          type: string
          format: uuid
          x-go-type: string
        answer:
          type: string
          example: 'apple'
        token:
          type: string
          format: uuid
          x-go-type: string
    SubmitAnswerResponse:
      type: object
      required:
        - correct
        - scrambledWord
        - scrambledWordDefinition
        - correctAttemptTimestamps
        - token
        - score
        - attempts
        - remainingSeconds
      properties:
        correct:
          type: boolean
        scrambledWord:
          type: string
          example: 'lpeap'
        scrambledWordDefinition:
          type: string
          example: 'A fruit that is often red or green.'
        correctAttemptTimestamps:
          type: array
          items:
            type: array
            items:
              type: string
              format: date-time
        token:
          type: string
          format: uuid
          x-go-type: string
        score:
          type: integer
          format: int32
          example: 10
        attempts:
          type: integer
          format: int32
          example: 2
        remainingSeconds:
          type: integer
          format: int32
          example: 45
    GetLeaderboardResponse:
      type: object
      required:
        - entries
        - total
      properties:
        entries:
          type: array
          items:
            type: object
            $ref: '#/components/schemas/LeaderboardEntry'
        total:
          type: integer
          format: int64
    LeaderboardEntry:
      type: object
      required:
        - id
        - rank
        - score
        - totalAttempts
        - accuracy
        - timestamp
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        isSessionSameAsCurrentUser:
          type: boolean
          example: false
        rank:
          type: integer
          format: int32
          example: 1
        score:
          type: integer
          format: int32
          example: 100
        totalAttempts:
          type: integer
          format: int32
          example: 5
        accuracy:
          type: number
          format: float
          example: 0.8
        timestamp:
          type: integer
          format: int64
          example: 1627847284