# packages/server/internal/api/gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// GetLeaderboardResponse defines model for GetLeaderboardResponse.
type GetLeaderboardResponse struct {
	Entries []LeaderboardEntry `json:"entries"`
	Total   int64              `json:"total"`
}

// LeaderboardEntry defines model for LeaderboardEntry.
type LeaderboardEntry struct {
	Accuracy                   float32 `json:"accuracy"`
	Id                         string  `json:"id"`
	IsSessionSameAsCurrentUser *bool   `json:"isSessionSameAsCurrentUser,omitempty"`
	Rank                       int32   `json:"rank"`
	Score                      int32   `json:"score"`
	Timestamp                  int64   `json:"timestamp"`
	TotalAttempts              int32   `json:"totalAttempts"`
}

// StartGameRequest defines model for StartGameRequest.
type StartGameRequest struct {
	AutoVoice *bool    `json:"autoVoice,omitempty"`
	ItemsUsed []string `json:"itemsUsed"`
}

// StartGameResponse defines model for StartGameResponse.
type StartGameResponse struct {
	AutoVoice               bool   `json:"autoVoice"`
	RemainingSeconds        int32  `json:"remainingSeconds"`
	ScrambledWord           string `json:"scrambledWord"`
	ScrambledWordDefinition string `json:"scrambledWordDefinition"`
	SessionId               string `json:"sessionId"`
	Token                   string `json:"token"`
}

// SubmitAnswerRequest defines model for SubmitAnswerRequest.
type SubmitAnswerRequest struct {
	Answer    string `json:"answer"`
	SessionId string `json:"sessionId"`
	Token     string `json:"token"`
}

// SubmitAnswerResponse defines model for SubmitAnswerResponse.
type SubmitAnswerResponse struct {
	Attempts                 int32         `json:"attempts"`
	Correct                  bool          `json:"correct"`
	CorrectAttemptTimestamps [][]time.Time `json:"correctAttemptTimestamps"`
	RemainingSeconds         int32         `json:"remainingSeconds"`
	Score                    int32         `json:"score"`
	ScrambledWord            string        `json:"scrambledWord"`
	ScrambledWordDefinition  string        `json:"scrambledWordDefinition"`
	Token                    string        `json:"token"`
}

// GetApiV1LeaderboardParams defines parameters for GetApiV1Leaderboard.
type GetApiV1LeaderboardParams struct {
	Page  *int `form:"page,omitempty" json:"page,omitempty"`
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// PostApiV1GameStartJSONRequestBody defines body for PostApiV1GameStart for application/json ContentType.
type PostApiV1GameStartJSONRequestBody = StartGameRequest

// PostApiV1GameSubmitJSONRequestBody defines body for PostApiV1GameSubmit for application/json ContentType.
type PostApiV1GameSubmitJSONRequestBody = SubmitAnswerRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /api/v1/game/continue)
	PostApiV1GameContinue(c *gin.Context)

	// (GET /api/v1/game/sse/{sessionId})
	GetApiV1GameSseSessionId(c *gin.Context, sessionId string)

	// (POST /api/v1/game/start)
	PostApiV1GameStart(c *gin.Context)

	// (POST /api/v1/game/submit)
	PostApiV1GameSubmit(c *gin.Context)

	// (GET /api/v1/leaderboard)
	GetApiV1Leaderboard(c *gin.Context, params GetApiV1LeaderboardParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// PostApiV1GameContinue operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1GameContinue(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1GameContinue(c)
}

// GetApiV1GameSseSessionId operation middleware
func (siw *ServerInterfaceWrapper) GetApiV1GameSseSessionId(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "sessionId" -------------
	var sessionId string

	err = runtime.BindStyledParameterWithOptions("simple", "sessionId", c.Param("sessionId"), &sessionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sessionId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiV1GameSseSessionId(c, sessionId)
}

// PostApiV1GameStart operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1GameStart(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1GameStart(c)
}

// PostApiV1GameSubmit operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1GameSubmit(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiV1GameSubmit(c)
}

// GetApiV1Leaderboard operation middleware
func (siw *ServerInterfaceWrapper) GetApiV1Leaderboard(c *gin.Context) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetApiV1LeaderboardParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page", c.Request.URL.Query(), &params.Page, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", c.Request.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiV1Leaderboard(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/api/v1/game/continue", wrapper.PostApiV1GameContinue)
	router.GET(options.BaseURL+"/api/v1/game/sse/:sessionId", wrapper.GetApiV1GameSseSessionId)
	router.POST(options.BaseURL+"/api/v1/game/start", wrapper.PostApiV1GameStart)
	router.POST(options.BaseURL+"/api/v1/game/submit", wrapper.PostApiV1GameSubmit)
	router.GET(options.BaseURL+"/api/v1/leaderboard", wrapper.GetApiV1Leaderboard)
}